theme: evict ES from number_of_records filter
