--- inspirehep/modules/theme/templates/inspirehep_theme/search/collection_jobs.html
@@ -1,35 +1,0 @@
-{#
-# This file is part of INSPIRE.
-# Copyright (C) 2015 CERN.
-#
-# INSPIRE is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License as
-# published by the Free Software Foundation; either version 2 of the
-# License, or (at your option) any later version.
-#
-# INSPIRE is distributed in the hope that it will be useful, but
-# WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-# General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with INSPIRE; if not, write to the Free Software Foundation, Inc.,
-# 59 Temple Place, Suite 330, Boston, MA 02111-1307, USA.
-#}
-
-{%- extends "inspirehep_theme/search/collection_basic.html" -%}
-
-{% block body %}
-  <div id="collection-search-bar" class="jobs-collection-search-bar">
-    <div class="title" id="jobs-search-bar">
-      Search {{ "Jobs" | number_of_records }} jobs
-    </div>
-    {% block search_bar %}
-      {{ super() }}
-    {% endblock search_bar %}
-    <div class="subtitle">
-      Example query:
-      <a href="/search?p=particle+physics&cc=jobs">particle physics</a>
-    </div>
-  </div>
-{% endblock %}
